+++ saleor/graphql/payment/mutations/transaction/transaction_request_action.py
@@ -34,7 +34,7 @@ from ....plugins.dataloaders import get_plugin_manager_promise
 from ....site.dataloaders import get_site_promise
 from ...enums import TransactionActionEnum
 from ...types import TransactionItem
-from ...utils import validate_and_resolve_refund_reason_context
+from ...utils import validate_reason_reference_context
 from .utils import create_transaction_event_requested, get_transaction_item
 
 if TYPE_CHECKING:
@@ -212,28 +212,28 @@ class TransactionRequestAction(BaseMutation):
 
         site = get_site_promise(info.context).get()
 
-        refund_reason_context = validate_and_resolve_refund_reason_context(
+        refund_reason_reference_type = site.settings.refund_reason_reference_type
+        should_apply = validate_reason_reference_context(
             reason_reference_id=reason_reference_id,
             requestor_is_user=bool(requestor_is_user),
-            refund_reference_field_name="refund_reason_reference",
+            reason_reference_field_name="refund_reason_reference",
             error_code_enum=TransactionRequestActionErrorCode,
-            site_settings=site.settings,
+            reason_reference_type=refund_reason_reference_type,
         )
-
-        refund_reason_reference_type = refund_reason_context[
-            "refund_reason_reference_type"
-        ]
+        refund_reason_reference_type_pk = (
+            refund_reason_reference_type.pk if refund_reason_reference_type else None
+        )
 
         reason_reference_instance: Page | None = None
 
-        if refund_reason_context["should_apply"]:
+        if should_apply and refund_reason_reference_type_pk:
             try:
                 reason_reference_pk = cls.get_global_id_or_error(
                     str(reason_reference_id), only_type="Page", field="reason_reference"
                 )
 
                 reason_reference_instance = Page.objects.get(
-                    pk=reason_reference_pk, page_type=refund_reason_reference_type.pk
+                    pk=reason_reference_pk, page_type=refund_reason_reference_type_pk
                 )
 
             except (Page.DoesNotExist, ValueError):